Reorganize files, remove unused components, and refactor button assets
