+++ src/components/common/RadioButton.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
 import { View } from 'react-native';
 import {
-  LeftAlignContainer,
   RadioButtonFill,
   RadioButtonSelected,
   RadioButtonUnselected,
-  RightAlignContainer,
   RowWithBottomMargin,
-  SelectableOneLine,
 } from './styles';
 import { Body1Text } from '../../../assets/Fonts';
+import {
+  LeftAlignContainer,
+  RightAlignContainer,
+  SelectableRow,
+} from '../../../assets/Components';
 
 type RadioButtonProps = {
   data: string[];
@@ -26,7 +28,7 @@ export default function RadioButtons({
     <View>
       {data.map((item, index) => (
         <RowWithBottomMargin key={item}>
-          <SelectableOneLine onPress={() => setSelected(index)}>
+          <SelectableRow onPress={() => setSelected(index)}>
             <LeftAlignContainer>
               <Body1Text>{item}</Body1Text>
             </LeftAlignContainer>
@@ -39,7 +41,7 @@ export default function RadioButtons({
                 <RadioButtonUnselected />
               )}
             </RightAlignContainer>
-          </SelectableOneLine>
+          </SelectableRow>
         </RowWithBottomMargin>
       ))}
     </View>
