+++ src/components/transactions/VoucherCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Body1Text } from '../../../assets/Fonts';
 import { formatValueForDisplay } from '../../utils/displayUtils';
 import { LeftAlignColumn, H4Subheading } from './styles';
-import { Row } from '../../../assets/Components';
+import { Card } from '../../../assets/Components';
 import { RightAlignContainer } from '../common/styles';
 
 type VoucherCardProps = {
@@ -11,13 +11,13 @@ type VoucherCardProps = {
 };
 export default function VoucherCard({ serialNumber, value }: VoucherCardProps) {
   return (
-    <Row>
+    <Card>
       <LeftAlignColumn>
         <Body1Text>SN {serialNumber}</Body1Text>
       </LeftAlignColumn>
       <RightAlignContainer style={{ marginRight: 29 }}>
         <H4Subheading>${formatValueForDisplay(value)}</H4Subheading>
       </RightAlignContainer>
-    </Row>
+    </Card>
   );
 }
